refactor(player-property): move calculateUsedLevel from PlayerLevelHelper
calculateUsedLevel 本质是「对 properties 累加」，归属感是 PropertyService 而非
积分换算 helper。搬迁后 PlayerLevelHelper 只剩积分↔等级换算，边界更清晰。

- Add static PlayerPropertyService.calculateUsedLevel(PlayerPropertyItemDto[])
- Remove from PlayerLevelHelper
- Move corresponding spec to player-property.service.spec.ts
