+++ api/src/player/helpers/player-level.helper.ts
@@ -75,18 +75,4 @@ export class PlayerLevelHelper {
   static getMemberLevelBuyPoint(point: number): number {
     return Math.floor(Math.sqrt(point / 25 + 380.25) - 19.5) + 1;
   }
-
-  /**
-   * 计算已使用的等级（从玩家属性中累加）
-   *
-   * TODO[migrate-usedlevel]: Step 3 切换读取路径后，此方法只剩 PlayerPropertyService.upgrade
-   *   一个调用方（用于计算写入 player.usedLevel 的绝对值）。届时把实现内联或下沉为
-   *   PlayerPropertyService 的 private 方法，并从此 helper 移除。
-   *
-   * @param properties 任何含 level 字段的对象数组（PropertyItem / PlayerPropertyItemDto 等）
-   * @returns 已使用的等级总数
-   */
-  static calculateUsedLevel(properties: Array<{ level: number }>): number {
-    return properties.reduce((prev, curr) => prev + curr.level, 0);
-  }
 }
